Now drones leave when adverse climate

+++ PracticaSD/Nucleo/AD_Engine.py
@@ -59,6 +59,7 @@ class AD_Engine:
         self.dronesDesactivados = []
         self.detener = False
         self.detener_por_clima = False
+        self.en_base_por_clima = False
 
     #Limpia la terminal para mejor legibilidad
     def clear_terminal(self):
@@ -136,7 +137,7 @@ class AD_Engine:
             inicio = time.time()
 
             #Escucha las posiciones de los drones mientras la figura no este completada o no se haya detenido el programa
-            while not self.detener and not self.detener_por_clima:
+            while not self.detener and not self.en_base_por_clima:
                 mensaje = consumidor.poll(0.1)
                 if mensaje is not None:
                     if mensaje.error():
@@ -205,6 +206,7 @@ class AD_Engine:
         try:
             topic = "destino"
             #Envio los destinos de los que se encuentran aqui
+
             for drone in self.dronesFinales:
                 pos = drone[1]
                 mensaje = f"{str(drone[0])} {str(pos[0])} {str(pos[1])}"
@@ -226,7 +228,10 @@ class AD_Engine:
             self.clear_terminal()
 
             if self.figura_completada():
-                mensaje = "*********************************************Figura Completada******************************************************" + "\n" + mensaje
+                if self.detener_por_clima:
+                    mensaje = "**************************CONDICIONES CLIMATICAS ADVERSAS. ESPECTACULO FINALIZADO***********************************" + "\n" + mensaje
+                else:
+                    mensaje = "*********************************************Figura Completada******************************************************" + "\n" + mensaje
 
             print(mensaje)
             print(self.dronesActuales)
@@ -243,7 +248,7 @@ class AD_Engine:
         try:
             topic = "posiciones"
             consumidor.subscribe(topics=[topic])
-            while not self.figura_completada() and not self.detener and not self.detener_por_clima:
+            while not self.figura_completada() and not self.detener and not self.en_base_por_clima:
                 mensaje = consumidor.poll(0.1)
                 if mensaje is not None:
                     if mensaje.error():
@@ -314,16 +319,29 @@ class AD_Engine:
                 pass
 
         # Eliminar el archivo JSON
-        try:
-            os.remove('Figuras.json')
-            print(f"El archivo {'Figuras.json'} fue eliminado con éxito.")
-        except OSError as e:
-            print(f"No se pudo eliminar el archivo {'Figuras.json'}: {e}")
+        if self.detener_por_clima == False:
+            try:
+                os.remove('Figuras.json')
+                print(f"El archivo {'Figuras.json'} fue eliminado con éxito.")
+            except OSError as e:
+                print(f"No se pudo eliminar el archivo {'Figuras.json'}: {e}")
 
+        return True
+
+    #Veo si todos los drones estan en la base por el clima
+    def retirada_clima(self):
+        for index in range(len(self.dronesActuales)):
+            if self.dronesActuales[index][1] != [-1,-1]:
+                return False
+        
         return True
 
     #Veo si la figura está completada o no
     def figura_completada(self):
+        if self.detener_por_clima == True and self.retirada_clima() == True:
+            en_base_por_clima = True
+            return True
+
         return self.dronesActuales == self.dronesFinales
 
     #Operaciones con sockets
@@ -348,6 +366,9 @@ class AD_Engine:
     def stop_clima(self):
         self.detener_por_clima = True
 
+        for index in range(len(self.dronesFinales)):
+            self.dronesFinales[index][1] = [-1,-1]
+
     #Se detiene por un mal funcionamiento
     def stop(self):
         self.detener = True
@@ -358,7 +379,7 @@ class AD_Engine:
     def start(self, productor_destinos, productor_mapa, consumidor_posiciones, consumidor_activos):
         hay_figura = self.leer_figuras()    #Lee las figuras del fichero json
         try:
-            while hay_figura and self.detener == False and self.detener_por_clima == False:
+            while hay_figura and self.detener == False and self.en_base_por_clima == False:
                 if not self.figuras:
                     hay_figura = False
                 else:
@@ -399,7 +420,7 @@ class AD_Engine:
 
                     #Mientras que no se haya completado la figura o no se haya detenido por ninguna causa, envia los destinos finales para los drones
                     #y el mapa actual a los drones.
-                    while not self.figura_completada() and self.detener == False and not self.detener_por_clima:
+                    while not self.figura_completada() and self.detener == False and not self.en_base_por_clima:
                         self.enviar_por_kafka_destinos(productor_destinos) #Envio los destinos a los drones
                         self.enviar_mapa(productor_mapa) #Envio el mapa una ultima vez porque sale del bucle antes de imprimir y enviar el ultimo mensaje
 
@@ -413,6 +434,10 @@ class AD_Engine:
 
                     self.enviar_mapa(productor_mapa) #Envio el mapa una ultima vez porque sale del bucle antes de imprimir y enviar el ultimo mensaje
                     del self.figuras[0] #Elimina la figura debido a que se ha completado
+
+                    #Si se ha parado por el clima se detiene la ejecucion
+                    if self.detener_por_clima == True and self.en_base_por_clima == True:
+                        break 
 
                     #Espero a que no exista ya el mismo hilo.
                     if dronesActivos.is_alive() == False:
@@ -457,11 +482,12 @@ class AD_Engine:
                             return
                     else:
                         hay_figura = True
+
             #Si la ejecución se ha detenido por el clima sale el mensaje de finalización
             if self.detener_por_clima == True:
                 print("CONDICIONES CLIMATICAS ADVERSAS.ESPECTACULO FINALIZADO")
-                if posicionesDrones.is_alive():
-                    posicionesDrones.detener()
+                if escucharPosiciones.is_alive():
+                    escucharPosiciones.detener()
 
             sys.exit(0)
         except Exception as e:
